Introduce Strategies directory, move code
