+++ packages/integrations-hcl/strategies/default/variables.hcl.ts
